Made amounts whole numbers

+++ database/models.py
@@ -25,7 +25,7 @@ class Order(Base):
     id = Column(Integer, primary_key=True)
     user_id = Column(Integer, ForeignKey('User.id'), nullable=False)
     order_type = Column(String(80), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Integer, nullable=False)
     interest_rate = Column(Float, nullable=False)
 
     def __repr__(self):
@@ -37,7 +37,7 @@ class Contract(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     borrower_id = Column(Integer, ForeignKey('Order.id'), nullable=False)
     lender_id = Column(Integer, ForeignKey('Order.id'), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Integer, nullable=False)
     interest_rate = Column(Float, nullable=False)
     date_created = Column(Date(), default=datetime.utcnow)
